fix: split the Android application into its own module
Completes the migration off the deprecated KMP + AGP pairing. The
compatibility shim is removed in AGP 10 (H2 2026) and Gradle states
the required shape outright: move the usage of com.android.application
into a separate subproject. There is no in-place option — the new
com.android.kotlin.multiplatform.library plugin is library-only.

Almost no source moved. An Android library can declare activities,
services and receivers in its own manifest and have them merge into
the app, so all 430-odd files stay in :composeApp, which is now a KMP
library and still owns the desktop distribution and the iOS
ComposeApp framework. Because the module name is unchanged, nomac's
Xcode project needs no edit. The new :androidApp carries only what is
genuinely application-level — applicationId, signing, ABI splits, R8,
packaging, lint — and has no Kotlin source of its own.

Three things that had to be discovered rather than read off:

Android resources are opt-in under the new plugin, unlike
com.android.library. Without androidResources { enable = true } no R
class is generated and the launcher icon, theme and shortcut
resources stop resolving.

The plugin must be registered at the root with `apply false`, or
per-module resolution fails with "already on the classpath with an
unknown version" — it arrives transitively through AGP.

org.jetbrains.kotlin.android must NOT be applied to :androidApp. Once
the bypass flags come off, AGP's built-in Kotlin owns compilation and
applying the standalone plugin is an error.

Both bypass flags are gone: android.builtInKotlin=false and
android.newDsl=false were a stopgap whose own comment said "until the
migration is done", and which had themselves become deprecated.

The version moved to gradle.properties because three consumers need
it — :composeApp for TalonBuild and the desktop package, :androidApp
for versionCode/versionName, and release.yml to name artifacts. That
also retires the regex over build.gradle.kts that had to avoid the
defaultConfig assignments, a subtlety that once surfaced as a bogus
tag mismatch.

Release plumbing updated and proven rather than assumed: a real
:androidApp:assembleRelease produces androidApp-{universal,arm64-v8a,
armeabi-v7a,x86_64}-release.apk with R8 and splits intact, which are
exactly the names release.yml now looks for. Desktop deb/dmg/msi
still build from :composeApp and are untouched. 1307 tests, every
compile target, and the icon keep-list audit all pass.
