+++ gradle.properties
@@ -21,7 +21,11 @@ ksp.useKSP2=true
 android.nonTransitiveRClass=true
 android.enableR8.fullMode=true
 
-# AGP 9 dropped com.android.application × KMP-plugin compat; bypass until
-# the com.android.kotlin.multiplatform.library migration is done.
-android.builtInKotlin=false
-android.newDsl=false
+
+# The app version, in one place because two modules and CI all need it:
+# :composeApp bakes it into TalonBuild and the desktop package,
+# :androidApp uses it for versionCode/versionName, and release.yml
+# reads it to name artifacts. It lived in composeApp/build.gradle.kts
+# until the Android application moved to its own module.
+talon.versionCode=223
+talon.versionName=0.16.0-rc25
